Implements the object_property_is_true function
This eliminates the need to manually get and check the property
value every time the program wants to know whether the value of the
property is considered truthy or not and also allows for the result
to default to true or false as the caller wishes if the property is
not defined.

This commit also includes changes to the verblib that make use of
the new function.